# Add some optimizations to utf8 decoding