This does four things:

1. adds Chunk.Queue.startsWith so we can be a bit more precise when checking for utf8 byte order mark
2. be miserly in internal methods with allocations and use null instead of Option (avoid allocation of Some on a loop).
3. use a Builder to avoid having to return a tuple and to reverse a list in an internal method.
4. avoid fold on an internal loop and just write out the while loop.