@johnynek 2020-07-11
```
Benchmark results:

on main:
[info] Benchmark                  (asciiStringSize)   Mode  Cnt       Score       Error  Units
[info] TextBenchmark.asciiDecode                128  thrpt    6  195334.730 ± 18599.880  ops/s
[info] TextBenchmark.asciiDecode               1024  thrpt    6  153746.264 ± 61576.197  ops/s
[info] TextBenchmark.asciiDecode               4096  thrpt    6  118761.218 ±  4695.784  ops/s

on this branch:
[info] Benchmark                  (asciiStringSize)   Mode  Cnt       Score      Error  Units
[info] TextBenchmark.asciiDecode                128  thrpt    6  218236.471 ± 1145.261  ops/s
[info] TextBenchmark.asciiDecode               1024  thrpt    6  192552.031 ± 1689.683  ops/s
[info] TextBenchmark.asciiDecode               4096  thrpt    6  146657.477 ±  589.035  ops/s
```

So this is 11% to 25% faster depending on the size, for this simple benchmark.